+++ internal/common/cleanup/on_timeout.go
@@ -3,7 +3,14 @@ package cleanup
 import (
 	"context"
 	"errors"
+	"strings"
 	"time"
+
+	"github.com/hashicorp/terraform-plugin-framework/diag"
+)
+
+const (
+	CleanupWarning = "Failed to create resource. Will run cleanup due to the operation timing out"
 )
 
 // OnTimeout creates a new context with a timeout and a deferred function that will run `cleanup` when the context hit the timeout (no timeout=no-op).
@@ -19,11 +26,27 @@ func OnTimeout(ctx context.Context, timeout time.Duration, warnDiags func(string
 		if !errors.Is(outCtx.Err(), context.DeadlineExceeded) {
 			return
 		}
-		cleanupWarning := "Failed to create a cluster, will run cleanup due to timeout reached"
-		warnDiags(cleanupWarning, warningDetail)
+		warnDiags(CleanupWarning, warningDetail)
 		newContext := context.Background() // Create a new context for cleanup as the old context is expired
 		if err := cleanup(newContext); err != nil {
 			warnDiags("Error during cleanup", warningDetail+" error="+err.Error())
+		}
+	}
+}
+
+const (
+	contextDeadlineExceeded = "context deadline exceeded"
+	TimeoutReachedPrefix    = "Timeout reached after "
+)
+
+func ReplaceContextDeadlineExceededDiags(diags *diag.Diagnostics, duration time.Duration) {
+	for i := range len(*diags) {
+		d := (*diags)[i]
+		if strings.Contains(d.Detail(), contextDeadlineExceeded) {
+			(*diags)[i] = diag.NewErrorDiagnostic(
+				d.Summary(),
+				strings.ReplaceAll(d.Detail(), contextDeadlineExceeded, TimeoutReachedPrefix+duration.String()),
+			)
 		}
 	}
 }