+++ internal/service/mongodbemployeeaccessgrant/model.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/hashicorp/terraform-plugin-framework/types"
 	"github.com/mongodb/terraform-provider-mongodbatlas/internal/common/conversion"
-	"go.mongodb.org/atlas-sdk/v20250312003/admin"
+	"go.mongodb.org/atlas-sdk/v20250312004/admin"
 )
 
 func NewTFModel(projectID, clusterName string, apiResp *admin.EmployeeAccessGrant) *TFModel {