+++ internal/service/organization/resource_organization_test.go
@@ -9,7 +9,7 @@ import (
 	"strings"
 	"testing"
 
-	"go.mongodb.org/atlas-sdk/v20250312003/admin"
+	"go.mongodb.org/atlas-sdk/v20250312004/admin"
 
 	"github.com/hashicorp/terraform-plugin-testing/helper/resource"
 	"github.com/hashicorp/terraform-plugin-testing/terraform"
@@ -168,7 +168,7 @@ func TestAccConfigDSOrganization_noAccessShouldFail(t *testing.T) {
 		Steps: []resource.TestStep{
 			{
 				Config:      configWithPluralDS("555") + acc.ConfigOrgMemberProvider(),
-				ExpectError: regexp.MustCompile("error getting organization settings .*"),
+				ExpectError: regexp.MustCompile("error getting organizations information:"),
 			},
 		},
 	})